auth-service/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
passlib[argon2]
python-jose
pydantic
